Add support for VP8, VP9 ,H265 and fix problème with bitrate because framerate was set to 1 causing the SDK to take a huge bitrate
