+++ sensorhub-driver-android/src/main/java/org/sensorhub/impl/sensor/android/video/AndroidCameraOutputH265.java
@@ -1,0 +1,62 @@
+/***************************** BEGIN LICENSE BLOCK ***************************
+
+The contents of this file are subject to the Mozilla Public License, v. 2.0.
+If a copy of the MPL was not distributed with this file, You can obtain one
+at http://mozilla.org/MPL/2.0/.
+
+Software distributed under the License is distributed on an "AS IS" basis,
+WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License
+for the specific language governing rights and limitations under the License.
+ 
+Copyright (C) 2012-2015 Sensia Software LLC. All Rights Reserved.
+ 
+******************************* END LICENSE BLOCK ***************************/
+
+package org.sensorhub.impl.sensor.android.video;
+
+import android.graphics.SurfaceTexture;
+import android.media.MediaCodec;
+import android.media.MediaCodecInfo;
+import android.media.MediaFormat;
+
+import org.sensorhub.api.sensor.SensorException;
+import org.sensorhub.impl.sensor.android.AndroidSensorsDriver;
+
+
+/**
+ * <p>
+ * Implementation of data interface for Android cameras using legacy Camera API.
+ * This will encode the video frames as a raw H264 stream (i.e. NAL units).
+ * </p>
+ *
+ * @author Alex Robin <alex.robin@sensiasoftware.com>
+ * @since June 11, 2015
+ */
+@SuppressWarnings("deprecation")
+public class AndroidCameraOutputH265 extends AndroidCameraOutput {
+
+    public AndroidCameraOutputH265(AndroidSensorsDriver parentModule, int cameraId, SurfaceTexture previewTexture) throws SensorException {
+        super(parentModule,cameraId,previewTexture,"camera" + cameraId + "_H265");
+    }
+
+
+    @Override
+    protected void initCodec() throws SensorException {
+        try {
+            final String videoCodec = MediaFormat.MIMETYPE_VIDEO_HEVC;
+            mCodec = MediaCodec.createEncoderByType(videoCodec);
+            MediaFormat mediaFormat = MediaFormat.createVideoFormat(videoCodec, imgWidth, imgHeight);
+            mediaFormat.setInteger(MediaFormat.KEY_BIT_RATE, 5 * 1000 * 1000 ); //2MB/s
+            mediaFormat.setInteger(MediaFormat.KEY_BITRATE_MODE, MediaCodecInfo.EncoderCapabilities.BITRATE_MODE_VBR);
+            mediaFormat.setInteger(MediaFormat.KEY_FRAME_RATE, 25);
+            mediaFormat.setInteger(MediaFormat.KEY_COLOR_FORMAT, MediaCodecInfo.CodecCapabilities.COLOR_FormatYUV420SemiPlanar);
+            mediaFormat.setInteger(MediaFormat.KEY_I_FRAME_INTERVAL, 1);
+            mCodec.configure(mediaFormat, null, null, MediaCodec.CONFIGURE_FLAG_ENCODE);
+            log.debug("MediaCodec initialized");
+        }
+        catch (Exception e)
+        {
+            throw new SensorException("Cannot initialize codec " + mCodec.getName(), e);
+        }
+    }
+}
